refactor: create useCreateRoleLogic
