+++ src/components/CreateRolePanel/CreateRolePanel.tsx
@@ -1,58 +1,18 @@
-import { useState } from 'react';
 import { Briefcase } from 'lucide-react';
-import type { NotificationState } from '../../types/global';
 import NotificationMessages from '../NotificationMessages/NotificationMessages';
+import { useCreateRoleLogic } from '../../hooks/useCreateRoleLogic'; 
 
 export default function CreateRolePanel() {
-  const [name, setName] = useState('');
-  const [description, setDescription] = useState('');
-  const [isSubmitting, setIsSubmitting] = useState(false);
-  const [notification, setNotification] = useState<NotificationState | null>(null);
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setNotification(null);
-
-    if (!name.trim() || !description.trim()) {
-      setNotification({ status: 'error', title: 'Campos Inválidos', message: 'Por favor, preencha todos os campos.' });
-      return;
-    }
-
-    setIsSubmitting(true);
-    const token = localStorage.getItem('token');
-    if (!token) {
-      setNotification({ status: 'error', title: 'Erro de Autenticação', message: 'Por favor, faça login novamente.' });
-      setIsSubmitting(false);
-      return;
-    }
-
-    try {
-      const requestBody = {
-        name,
-        type: 'CARGO', 
-        description,
-      };
-
-      const response = await fetch('http://localhost:3000/api/roles', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json', 'Authorization': `Bearer ${token}` },
-        body: JSON.stringify(requestBody),
-      });
-
-      if (!response.ok) {
-        throw new Error(`Erro ${response.status}: Não foi possível criar o cargo.`);
-      }
-
-      setNotification({ status: 'success', title: 'Sucesso!', message: `O cargo "${name}" foi criado.` });
-      setName('');
-      setDescription('');
-
-    } catch (error) {
-      setNotification({ status: 'error', title: 'Falha na Criação', message: (error as Error).message });
-    } finally {
-      setIsSubmitting(false);
-    }
-  };
+  const {
+    name,
+    setName,
+    description,
+    setDescription,
+    isSubmitting,
+    notification,
+    setNotification,
+    handleSubmit,
+  } = useCreateRoleLogic();
 
   const inputStyles = `mt-1 block w-full rounded-md shadow-sm border border-gray-300 py-2 px-3 focus:outline-none focus:ring-2 focus:ring-orange-500 focus:border-transparent`;
 
@@ -71,9 +31,10 @@ export default function CreateRolePanel() {
               id="role-name"
               type="text"
               value={name}
-              onChange={e => setName(e.target.value)}
+              onChange={e => setName(e.target.value)} 
               className={inputStyles}
               placeholder="Ex: Designer de Produto"
+              disabled={isSubmitting}
             />
           </div>
           
@@ -82,10 +43,11 @@ export default function CreateRolePanel() {
             <textarea
               id="role-description"
               value={description}
-              onChange={e => setDescription(e.target.value)}
+              onChange={e => setDescription(e.target.value)} 
               rows={4}
               className={inputStyles}
               placeholder="Descreva as responsabilidades principais deste cargo."
+              disabled={isSubmitting}
             />
           </div>
           
@@ -107,7 +69,7 @@ export default function CreateRolePanel() {
           status={notification.status}
           title={notification.title}
           message={notification.message}
-          onClose={() => setNotification(null)}
+          onClose={() => setNotification(null)} 
         />
       )}
     </div>
